+++ mayan/apps/appearance/static/appearance/js/mayan_image.js
@@ -6,23 +6,24 @@ class MayanImage {
         this.load();
     }
 
-    static intialize () {
-        var app = this;
+    static intialize (options) {
+        this.options = options || {};
+        this.options.templateInvalidDocument = this.options.templateInvalidDocument || '<span>Error loading document image</span>';
 
-        this.fancybox = $().fancybox({
-            animationDuration : 300,
-            buttons : [
-                'fullScreen',
-                'close',
-            ],
-            selector: 'a.fancybox',
+        $().fancybox({
             afterShow: function (instance, current) {
                 $('a.a-caption').on('click', function(event) {
                     instance.close(true);
                 });
             },
+            animationEffect: 'fade',
+            animationDuration : 100,
+            buttons : [
+                'fullScreen',
+                'close',
+            ],
             infobar: true,
-
+            selector: 'a.fancybox'
         });
 
         $('img.lazy-load').lazyload({
@@ -42,46 +43,53 @@ class MayanImage {
 
         $('.lazy-load').one('load', function() {
             $(this).hide();
-            $(this).fadeIn(300);
+            $(this).show();
             $(this).siblings('.spinner-container').remove();
             $(this).removeClass('lazy-load pull-left');
             clearTimeout(MayanImage.timer);
-            MayanImage.timer = setTimeout(MayanImage.timerFunction, 100);
+            MayanImage.timer = setTimeout(MayanImage.timerFunction, 250);
         });
 
         $('.lazy-load-carousel').one('load', function() {
             $(this).hide();
-            $(this).fadeIn(300);
+            $(this).show();
             $(this).siblings('.spinner-container').remove();
             $(this).removeClass('lazy-load-carousel pull-left');
         });
     }
 
     static timerFunction () {
-        $.fn.matchHeight._maintainScroll = true;
         $.fn.matchHeight._update();
     }
 
     load () {
         var self = this;
         var container = this.element.parent().parent().parent();
+        var dataURL = this.element.attr('data-url');
 
-        this.element.attr('src', this.element.attr('data-url'));
-        this.element.on('error', function() {
-            // Check the .complete property to see if it is a real error
-            // or it was a cached image
-            if (this.complete === false) {
-                // It is a cached image, set the src attribute to trigger
-                // it's display.
-                this.src = this.src;
-            } else {
-                container.html(MayanImage.templateInvalidDocument);
-            }
-        });
-
-        $.fn.matchHeight._maintainScroll = true;
+        if (dataURL === '') {
+            container.html(MayanImage.options.templateInvalidDocument);
+        } else {
+            this.element.attr('src', dataURL);
+            setTimeout(function () {
+                self.element.on('error', function () {
+                    // Check the .complete property to see if it is a real
+                    // error or it was a cached image
+                    if (this.complete === false) {
+                        // It is a cached image, set the src attribute to
+                        // trigger it's display.
+                        this.src = dataURL;
+                    } else {
+                        container.html(
+                            MayanImage.options.templateInvalidDocument
+                        );
+                    }
+                });
+            }, 1);
+        }
     };
 }
 
-MayanImage.templateInvalidDocument = $('#template-invalid-document').html();
 MayanImage.timer = setTimeout(null);
+
+$.fn.matchHeight._maintainScroll = true;